requirements.txt
opencc-python-reimplemented>=1.1.6